Added image and the credit for it.
"Academic" supports having an image/avatar. This has now been added in
the correct format (a square, .jpg image). However, there did not seem
to be a way to associate credit with this picture, so I edited the style
from the "Academic" theme for the "about" widget in order to support
this.